requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv